pyproject.toml
# pyproject.toml
[project]
name = "carrus"
version = "0.1.0"
description = "Modern macOS package manager with MDM integration"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "rich>=13.0.0",
    "PyYAML>=6.0.1",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "packaging>=23.0",
]

[project.scripts]
carrus = "carrus.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/carrus"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
]

ignore = [
    "E501",  # Line too long (temporarily)
    "S101",  # Use of assert (allow in tests)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"